# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "FormulaWeb QuickBooks Agent",
    "description": "Thin translator between the FormulaWeb ERP and QuickBooks Desktop. JSON in, JSON out.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost:8462/"
    }
  ],
  "paths": {
...
    "/journal-entries": {
      "get": {
        "tags": [
          "JournalEntries"
        ],
        "summary": "List journal entries, optionally only those modified at or after 'updatedSince'. 'maxReturned' caps the result; omit it to fetch all.",
        "operationId": "ListJournalEntries",
        "parameters": [
          {
            "name": "updatedSince",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "maxReturned",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JournalEntryDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "JournalEntries"
        ],
        "summary": "Post a journal entry to QuickBooks. Requires an Idempotency-Key header; replays are safe. QuickBooks enforces that debits equal credits.",
        "operationId": "CreateJournalEntry",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJournalEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalEntryDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/journal-entries/{txnId}": {
      "get": {
        "tags": [
          "JournalEntries"
        ],
        "summary": "Fetch a single journal entry by its QuickBooks TxnID.",
        "operationId": "GetJournalEntry",
        "parameters": [
          {
            "name": "txnId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalEntryDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateJournalEntryRequest": {
        "required": [
          "txnDate",
          "lines"
        ],
        "type": "object",
        "properties": {
          "txnDate": {
            "type": "string",
            "format": "date"
          },
          "refNumber": {
            "maxLength": 11,
            "type": [
              "null",
              "string"
            ]
          },
          "isAdjustment": {
            "type": "boolean"
          },
          "lines": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateJournalLine"
            }
          }
        }
      },
      "CreateJournalLine": {
        "required": [
          "type",
          "accountFullName",
          "amount"
        ],
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/JournalLineType"
          },
          "accountFullName": {
            "type": "string"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "memo": {
            "type": [
              "null",
              "string"
            ]
          },
          "entityFullName": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "JournalEntryDto": {
        "required": [
          "txnId",
          "editSequence",
          "txnDate",
          "lines"
        ],
        "type": "object",
        "properties": {
          "txnId": {
            "type": "string"
          },
          "editSequence": {
            "type": "string"
          },
          "txnNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "txnDate": {
            "type": "string",
            "format": "date"
          },
          "refNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "isAdjustment": {
            "type": "boolean"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JournalEntryLineDto"
            }
          },
          "timeCreated": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "timeModified": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "JournalEntryLineDto": {
        "required": [
          "type",
          "amount"
        ],
        "type": "object",
        "properties": {
          "txnLineId": {
            "type": [
              "null",
              "string"
            ]
          },
          "type": {
            "$ref": "#/components/schemas/JournalLineType"
          },
          "accountFullName": {
            "type": [
              "null",
              "string"
            ]
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "memo": {
            "type": [
              "null",
              "string"
            ]
          },
          "entityFullName": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "JournalLineType": {
        "enum": [
          "Debit",
          "Credit"
        ]
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      }
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "Shared secret for the internal ERP->agent network.",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
...
}